+++ metadata.rb
@@ -4,7 +4,7 @@ maintainer_email 'jeff@csd.uwo.ca'
 license          'Apache 2.0'
 description      'Installs/configures torquebox'
 long_description 'Installs/configures torquebox'
-version          '0.1.0'
+version          '0.1.1'
 
 supports 'centos', '>= 6.5'