+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+# 0.1.1
+
+* Fix namespacing issues in application provider
+* Fix jruby path errors in application provider
+
 # 0.1.0
 
 Initial release of torquebox